Yolo merging changes to master since they don't break any code
Broke apart CSS formatting for the table. Completely changed how I
generate the mine locations.

Working on assigning on click triggers. Essentially when a button is
pressed, it will now check the generated array of mines for existence,
and then act accordingly. Having timing issues with running the
listener — will continue hacking away.

Second part of this will involve “removing fog of war buttons” to
reflect outlines of other mines — or just go straight hardcore with it.

This will require the replacement of buttons with cells — which might
mean a complete re-write .